+++ crutch_core/src/main/java/sk/alival/crutch/README_core.md
@@ -39,3 +39,25 @@ Logs.init(
 	}  
 )
 ```
+
+# Core
+
+**Launcher as extensions for CoroutineScope**
+
+```kotlin
+launchIO
+launchDefault
+launchMain
+launchMainImmediate
+launchUnconfined
+```
+
+**Context switchers**
+
+```kotlin
+withMainImmediate
+withIO
+withDefault
+withMain
+withUnconfined
+```
