+++ crutch_core/src/main/java/sk/alival/crutch/coroutines/WithHelpers.kt
@@ -1,4 +1,4 @@
-package sk.alival.crutch.states.coroutineHelpers
+package sk.alival.crutch.coroutines
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
