+++ crutch_states/src/main/java/sk/alival/crutch/states/README_states.md
@@ -190,26 +190,6 @@ launchOnMain
 launchOnMainImmediate
 ```
 
-**Launcher as extensions for CoroutineScope**
-
-```kotlin
-launchIO
-launchDefault
-launchMain
-launchMainImmediate
-launchUnconfined
-```
-
-**Context switchers**
-
-```kotlin
-withMainImmediate
-withIO
-withDefault
-withMain
-withUnconfined
-```
-
 ## TODO
 
 - support for savedStateHandle
